requirements.txt
scikit-learn
matplotlib
seaborn
pydot
pqdm
dask
pandas-tfrecords
pyarrow
fastparquet
ipdb
tensorflow_addons
